src/test/java/game/GameTest.java
package game;

import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.mockito.Mockito;

import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;

public class GameTest {


    @Test
    @DisplayName("한 사이클 진행")
    public void oneCycle() {
        Car mySpringCar = Mockito.mock(Car.class);
        Car mySummerCar = Mockito.mock(Car.class);

        Game game = new Game(List.of(mySpringCar, mySummerCar));
        game.doACycle();

        Mockito.verify(mySpringCar).go();
        Mockito.verify(mySummerCar).go();
    }
}
